+++ json_annotation/lib/src/json_serializable.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:collection';
+
 class JsonSerializable {
   // TODO(kevmoo): document these fields
   final bool createFactory;
@@ -67,4 +69,56 @@ class JsonKey {
   ///
   /// Only required when the default behavior is not desired.
   const JsonKey({this.name, this.nullable, this.includeIfNull});
+}
+
+// TODO(kevmoo): Add documentation
+abstract class $JsonMapWrapper extends UnmodifiableMapBase<String, dynamic> {}
+
+Map<String, dynamic> $wrapMap<K, V>(
+        Map<K, V> source, dynamic converter(V key)) =>
+    new _MappingMap(source, converter);
+
+Map<String, dynamic> $wrapMapHandleNull<K, V>(
+        Map<K, V> source, dynamic converter(V key)) =>
+    source == null ? null : new _MappingMap(source, converter);
+
+List<dynamic> $wrapList<T>(List<T> source, dynamic converter(T key)) =>
+    new _MappingList(source, converter);
+
+List<dynamic> $wrapListHandleNull<T>(
+        List<T> source, dynamic converter(T key)) =>
+    source == null ? null : new _MappingList(source, converter);
+
+typedef dynamic _Convert<S>(S value);
+
+class _MappingList<S> extends ListBase<dynamic> {
+  final List<S> _source;
+  final _Convert<S> _converter;
+
+  _MappingList(this._source, this._converter);
+
+  @override
+  dynamic operator [](int index) => _converter(_source[index]);
+
+  @override
+  operator []=(int index, dynamic value) => throw new UnsupportedError('');
+
+  @override
+  int get length => _source.length;
+
+  @override
+  set length(int value) => throw new UnsupportedError('');
+}
+
+class _MappingMap<K, V> extends UnmodifiableMapBase<String, dynamic> {
+  final Map<K, V> _source;
+  final _Convert<V> _converter;
+
+  _MappingMap(this._source, this._converter);
+
+  @override
+  Iterable<String> get keys => _source.keys.map((k) => k as String);
+
+  @override
+  dynamic operator [](Object key) => _converter(_source[key]);
 }