+++ json_annotation/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.2
+
+* Added a helper class – `$JsonMapWrapper` – and helper functions – `$wrapMap`, 
+  `$wrapMapHandleNull`, `$wrapList`, and `$wrapListHandleNull` – to support
+  the `useWrappers` option added to `JsonSerializableGenerator` in `v0.3.0` of
+  `package:json_serializable`.
+
 ## 0.2.1
 
 * `JsonSerializable` class annotation